% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate-Functions.R
\name{agg_cors}
\alias{agg_cors}
\title{Aggregated factor correlations}
\usage{
agg_cors(kfa, type = "factor", flag = 0.9)
}
\arguments{
\item{kfa}{An object returned from \code{\link[kfa]{kfa}}}

\item{type}{currently ignored; \code{"factor"} (default) or \code{"observed"} variable correlations}

\item{flag}{threshold above which a factor correlation will be flagged}
}
\value{
a \code{data.frame} of mean factor correlations for each factor model and \code{vector} with count of folds with a flagged correlation
}
\description{
The factor correlations aggregated over k-folds
}
