% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate-Functions.R
\name{agg_fac_cor}
\alias{agg_fac_cor}
\title{Aggregated factor correlations}
\usage{
agg_fac_cor(kfa)
}
\arguments{
\item{kfa}{An object returned from \code{\link[kfa]{kfa}}}
}
\value{
\code{data.frame} of factor correlations for each factor model
}
\description{
The factor correlations aggregated over k-folds
}
