% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate-Functions.R
\name{agg_loadings}
\alias{agg_loadings}
\title{Aggregated factor loadings}
\usage{
agg_loadings(kfa)
}
\arguments{
\item{kfa}{An object returned from \code{\link[kfa]{kfa}}}
}
\value{
\code{data.frame} of mean factor loadings for each factor model
}
\description{
The factor loadings aggregated over k-folds
}
