% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate-Functions.R
\name{agg_model_fit}
\alias{agg_model_fit}
\title{Summary table of model fit}
\usage{
agg_model_fit(k.fits, index = "rmsea")
}
\arguments{
\item{k.fits}{An object returned from \code{\link[kfa]{k_model_fit}}}
}
\value{
For each factor quantity, a \code{list} containing \code{lavaan} syntax specifying the factor structure and the folds where the structure was identified
}
\description{
Summary table of average model fit from the k-folds
}
\details{
Only RMSEA is currently summarized
}
