% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate-Functions.R
\name{agg_param}
\alias{agg_param}
\title{Aggregate parameter and stability}
\usage{
agg_param(est, var, n, alpha = 0.05, output.var = FALSE)
}
\arguments{
\item{est}{vector of parameter estimates}

\item{var}{vector of variances for \code{est}}

\item{n}{number of observations per fold}

\item{alpha}{type I error rate for one-sided test}

\item{output.var}{include all variance components in the output}
}
\value{
a single row \code{data.frame}
}
\description{
Internal function for aggregating a parameter over k-folds following Rubin's (1987) rules and calculating the stability of the parameter.
}
\details{
The stability index is calculated as 1 - bv / tv where bv is the between fold variance and tv is the total pooled variance.
}
