% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate-Functions.R
\name{agg_reliability}
\alias{agg_reliability}
\title{Aggregated scale reliabilities}
\usage{
agg_reliability(kfa)
}
\arguments{
\item{kfa}{An object returned from \code{\link[kfa]{kfa}}}
}
\value{
\code{data.frame} of factor reliabilities for each factor model
}
\description{
The factor reliabilities aggregated over k-folds
}
