% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate-Functions.R
\name{appendix_prep}
\alias{appendix_prep}
\title{Internal appendix function}
\usage{
appendix_prep(fits, index, suffix)
}
\arguments{
\item{fits}{An element from the list object returned from \code{\link[kfa]{k_model_fit}} when \code{by.folds = FALSE}}

\item{index}{One or more fit indices to include in the appendix table. Default is \code{"all"} indices present in \code{fits}. The degrees of freedom are always reported.}

\item{suffix}{character to append to column names}
}
\value{
\code{data.frame} of model fit by fold, factor model, and fit index
}
\description{
Prepare model fit results for appendix table
}
