% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfa_report.R
\name{appendix_wrapper}
\alias{appendix_wrapper}
\title{Create appendix table}
\usage{
appendix_wrapper(appendix, mapping, border, digits)
}
\arguments{
\item{appendix}{object returned from \code{\link[kfa]{get_appendix}}}

\item{mapping}{a \code{data.frame} specifying header columns. See \code{\link[flextable]{set_header_footer_df}} for details.}

\item{border}{format of of horizontal borders. See \code{\link[flextable]{border_inner_h}} for details.}

\item{digits}{integer; number of decimal places to display in the report.}
}
\value{
a \code{flextable} object
}
\description{
Internal function for converting appendix from \code{data.frame} to \code{flextable}
}
