% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unused.R
\name{best_model}
\alias{best_model}
\title{Model with best fit}
\usage{
best_model(kfits, index = c("chisq", "cfi", "rmsea"))
}
\arguments{
\item{kfits}{An object returned from \code{\link[kfa]{k_model_fit}} when \code{by.folds = TRUE}}

\item{index}{One or more fit indices to summarize in the report. Default are "chisq", "cfi", and "rmsea".}
}
\value{
\code{data.frame} with fold in rows, index in columns, and factor model with best fit in the cells
}
\description{
Summary table of best fitting model in k-folds
}
