% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Syntax-Functions.R
\name{efa_cfa_syntax}
\alias{efa_cfa_syntax}
\alias{write_efa}
\title{Converts EFA results into CFA syntax}
\usage{
efa_cfa_syntax(
  loadings,
  simple = TRUE,
  threshold = NA,
  single.item = c("keep", "drop", "none")
)

write_efa(nf, vnames)
}
\arguments{
\item{loadings}{matrix of factor loadings}

\item{simple}{logical; Should the simple structure be returned (default)?
If \code{FALSE}, items can cross-load on multiple factors.}

\item{threshold}{numeric between 0 and 1 indicating the minimum (absolute) value
of the loading for an item on a factor.}

\item{single.item}{character indicating how single-item factors should be treated.
Use \code{"keep"} to keep them in the model when generating the CFA syntax, \code{"drop"}
to remove them, or \code{"none"} indicating the CFA syntax should not be generated for
this model and \code{""} will be returned.}

\item{nf}{integer; number of factors}

\item{vnames}{character vector; names of variables to include in the efa}
}
\value{
character. Use \code{cat()} to best examine the returned syntax.
}
\description{
Uses the loadings matrix (presumably from an EFA) to generate \code{lavaan} compatible CFA syntax.
}
