% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-kfa-data.R
\docType{data}
\name{example.kfa}
\alias{example.kfa}
\title{kfa results from simulated data example}
\format{
An object of class \code{"kfa"}, which is a four-element \code{list}:
\itemize{
\item \strong{cfas} \code{lavaan} CFA objects for each \emph{k} fold
\item \strong{cfa.syntax} syntax used to produce CFA objects
\item \strong{model.names} vector of names for CFA objects
\item \strong{efa.structures} all factor structures identified in the EFA
}
}
\usage{
data(example.kfa)
}
\description{
Simulated responses for 900 observations on 20 variables loading onto a 3 factor
structure (see example in \code{\link[kfa]{kfa}} documentation for model).
The simulated data was run through \code{\link[kfa]{kfa}} with the call
kfa(sim.data, k = 2, m = 3) which tested 1-, 2-, and 3-factor structures over 2 folds.
}
\examples{
data(example.kfa)
agg_cors(example.kfa)

}
\keyword{datasets}
