% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfa_report.R
\name{flextab_format}
\alias{flextab_format}
\title{Default flextable format}
\usage{
flextab_format(df, bold.type = "none", cut = 0.3, digits = 2)
}
\arguments{
\item{df}{a \code{data.frame}}

\item{bold.type}{character indicating table with a pre-specified bolding pattern. Not currently implemented.}

\item{cut}{numeric; standardized factor loadings below this value are bolded in tables.}

\item{digits}{integer; number of decimal places to display in the report.}
}
\value{
a \code{flextable} object
}
\description{
Internal function for formatting flextables
}
