% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate-Functions.R
\name{get_appendix}
\alias{get_appendix}
\title{Appendix of model fit}
\usage{
get_appendix(mfits, index = "all")
}
\arguments{
\item{mfits}{An object returned from \code{\link[kfa]{k_model_fit}} when \code{by.folds = FALSE}}

\item{index}{One or more fit indices to include in the appendix table. Default is \code{"all"} indices present in \code{mfits}. The degrees of freedom are always reported.}
}
\value{
\code{data.frame} of model fit by fold, factor model, and fit index
}
\description{
Comprehensive model fit table
}
