% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit-Functions.R
\name{index_available}
\alias{index_available}
\title{Available Fit Indices}
\usage{
index_available(models)
}
\arguments{
\item{models}{an object returned from \code{\link[kfa]{kfa}}}
}
\value{
character vector of index names
}
\description{
Shows the fit indices available from \code{\link[kfa]{kfa}} object to report in \code{\link[kfa]{kfa_report}}
}
\examples{
data(example.kfa)
index_available(example.kfa)

}
