% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_cfa.R
\name{k_cfa}
\alias{k_cfa}
\title{k-fold confirmatory factor analysis}
\usage{
k_cfa(syntax, variables, ordered, estimator, missing, ...)
}
\arguments{
\item{syntax}{A (named) list of \code{lavaan} syntax specifying the CFA models to run.}

\item{variables}{a \code{data.frame} of variables to factor analyze.}

\item{ordered}{passed to \code{lavaan} functions. See \code{\link[lavaan]{lavCor}}.}

\item{estimator}{passed to \code{lavaan} functions. See \code{\link[lavaan]{lavCor}}.}

\item{missing}{passed to \code{lavaan} functions. See \code{\link[lavaan]{lavCor}}.}

\item{...}{other arguments passed to \code{lavaan} functions. See \code{\link[lavaan]{lavOptions}}.}
}
\value{
A list of \code{lavaan} objects
}
\description{
Runs CFA for models identified from a previous EFA.
}
