% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_efa.R
\name{k_efa}
\alias{k_efa}
\title{k-fold exploratory factor analysis}
\usage{
k_efa(variables, m, rotation, ordered, estimator, missing, ...)
}
\arguments{
\item{variables}{dataframe of variables to factor analyze}

\item{m}{integer; maximum number of factors to extract.}

\item{rotation}{character (case-sensitive); any rotation method listed in
\code{\link[GPArotation]{rotations}} in the \code{GPArotation} package.}

\item{ordered}{passed to lavaan functions}

\item{estimator}{passed to lavaan functions}

\item{missing}{passed to lavaan functions}
}
\value{
A list containing \code{lavaan} compatible CFA syntax.
}
\description{
For each k fold, the function conducts a sequence of EFAs and
converts the resulting factor structure into \code{lavaan}
compatible CFA syntax.
}
