% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate-Functions.R
\name{k_model_fit}
\alias{k_model_fit}
\title{Extract model fit}
\usage{
k_model_fit(kfa)
}
\arguments{
\item{kfa}{An object returned from \code{\link[kfa]{kfa}}}
}
\value{
\code{data.frame} of average model fit for each factor model
}
\description{
Model fit indices extracted from k-folds
}
