% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfa.R
\name{kfa}
\alias{kfa}
\title{Conducts k-fold cross validation for factor analysis.}
\usage{
kfa(
  variables,
  k = NULL,
  rmsea0 = 0.05,
  rmseaA = 0.08,
  m = floor(ncol(items)/4),
  rotation = "oblimin",
  ordered = FALSE,
  estimator = "default",
  missing = "listwise",
  ...
)
}
\arguments{
\item{variables}{a \code{data.frame} of variables to factor analyze}

\item{k}{an integer between 2 and 10; number of folds in which to split the data. Default is NULL which determines k via power analysis.}

\item{rmsea0}{numeric; RMSEA under the null hypothesis for the power analysis.}

\item{rmseaA}{numeric; RMSEA under the alternative hypothesis for the power analysis.}

\item{m}{integer; maximum number of factors to extract. Default is 4 items per factor.}

\item{rotation}{character (case-sensitive); any rotation method listed in
\code{\link[GPArotation]{rotations}} in the \code{GPArotation} package. Default is "oblimin".}

\item{ordered}{logical; Should items be treated as ordinal and the
polychoric correlations used in the factor analysis? When \code{FALSE} (default)
the Pearson correlation matrix is used. A character vector of item names is
also accepted to prompt estimation of the polychoric correlation matrix.}

\item{estimator}{if \code{ordered = FALSE}, the default is "ML". If
\code{ordered = TRUE}, the default is "DWLS". See \code{\link[lavaan]{lavaan}} for other options.}

\item{missing}{default is "listwise". See \code{\link[lavaan]{lavaan}} for other options.}

\item{...}{other arguments passed to \code{lavaan} functions. See \code{\link[lavaan]{lavOptions}}.}
}
\value{
a \code{list} of \code{lists} with each outer element being a k-fold and each inner element a \code{lavaan} object
}
\description{
The function splits the data into k folds. For each fold,
EFAs are run on the training data and the simple structure each model
is transformed into \code{lavaan}-compatible CFA syntax. The CFAs are then run
on the test data.
}
\details{
Deciding an appropriate *m* can be difficult, but is consequential for both the
possible factor structures to examine and the computation time. To assist, researchers
can use \code{\link[parameters]{n_factors}} in the \code{parameters} package.
}
