% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_fa.R
\name{kfold_fa}
\alias{kfold_fa}
\title{Conducts k fold cross-validation for factor analysis.}
\usage{
kfold_fa(
  items,
  efa.method,
  rotation = "oblimin",
  k = NULL,
  rmsea0 = 0.05,
  rmseaA = 0.08,
  m = floor(ncol(items)/4),
  threshold = NA,
  ordered = FALSE,
  missing = "listwise",
  ...
)
}
\arguments{
\item{items}{dataframe of item responses}

\item{efa.method}{character; How should the EFA models be determined?
"sequence" sequentially runs 1 to *m* factor models.
"consensus" calls \code{\link[parameters]{n_factors}} which identifies the
optimal number of factors based the consensus of a variety of methods.}

\item{k}{integer; number of folds in which to split the data.
Default is NULL which determines k via power analysis.
The range for k is currently set to 2 to 10.}

\item{rmsea0}{numeric; RMSEA under the null hypothesis for the power analysis.}

\item{rmseaA}{numeric; RMSEA under the alternative hypothesis for the power analysis.}

\item{m}{integer; maximum number of factors to extract. Default is 4 items per factor.}

\item{ordered}{logical; Should items be treated as ordinal and the
polychoric correlations used in the factor analysis? When \code{FALSE} (default)
the Pearson correlation matrix is used. A character vector of item names is
also accepted to prompt estimation of the polychoric correlation matrix.}

\item{...}{arguments to pass to \code{lavaan}. These may include ...}
}
\description{
The function splits the data into k folds. For each fold,
EFAs are run on the training data and the simple structure each model
is transformed into \code{lavaan}-compatible CFA syntax. The CFAs are then run
on the test data with model comparison results summarized in a rmarkdown
produced report.
}
