% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Syntax-Functions.R
\name{match_structure}
\alias{match_structure}
\title{Match factor structures}
\usage{
match_structure(structures)
}
\arguments{
\item{structures}{list of \code{lavaan} syntax}
}
\value{
For each unique structure, a \code{list} containing \code{lavaan} syntax specifying the factor structure and the folds where the structure was identified
}
\description{
Internal function in model structures
}
