% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate-Functions.R
\name{model_flags}
\alias{model_flags}
\title{Flag model problems}
\usage{
model_flags(kfa, strux, loads, cors, rels)
}
\arguments{
\item{kfa}{An object returned from \code{\link[kfa]{kfa}}}

\item{strux}{An object returned from \code{\link[kfa]{model_structures}} when \code{which = "cfa"}}

\item{loads}{An object returned from \code{\link[kfa]{agg_loadings}}}

\item{cors}{An object returned from \code{\link[kfa]{agg_cors}}}

\item{rels}{An object returned from \code{\link[kfa]{agg_rels}}}
}
\value{
a \code{data.frame}
}
\description{
Internal function to create table of flag counts
}
