% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate-Functions.R
\name{model_structure}
\alias{model_structure}
\title{Unique factor structures}
\usage{
model_structure(kfa)
}
\arguments{
\item{kfa}{An object returned from \code{\link[kfa]{kfa}}}
}
\value{
A \code{list} containing \code{lavaan} syntax specifying the factor structure and the folds where the structure was identified
}
\description{
Extract unique factor structures across the k-folds
}
