% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_efa.R
\name{run_efa}
\alias{run_efa}
\title{Conducts exploratory factor analysis}
\usage{
run_efa(items, rotation, ordered, missing, ...)
}
\arguments{
\item{items}{dataframe of item responses}

\item{rotation}{character; any rotation method listed in \code{\link[GPArotation]{rotations}}
in the \code{GPArotation} package. Default is "oblimin".}

\item{ordered}{logical; Should variables be treated as ordinal and the polychoric correlations used in the factor analysis?
  Default is FALSE where all variables are considered numeric and the Pearson correlation matrix is calculated.
A character vector of variable names is also accepted to prompt estimation of the polychoric correlation matrix.}

\item{missing}{passed to lavaan functions}

\item{...}{other arguments to pass on to lavaan}
}
\description{
This function is intended for use on independent samples rather than integrated
with k-fold cross-validation.
}
