% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfa_report.R
\name{two_level_flex}
\alias{two_level_flex}
\title{Default flextable format for header with two levels}
\usage{
two_level_flex(flex, mapping, vert.cols, border)
}
\arguments{
\item{flex}{a \code{flextable} object}

\item{mapping}{a \code{data.frame} specifying header columns. See \code{\link[flextable]{set_header_footer_df}} for details.}

\item{vert.cols}{columns where level 1 and level 2 cells will be vertically merged. See \code{\link[flextable]{merge_v}} for details.}

\item{border}{format of of horizontal borders. See \code{\link[flextable]{border_inner_h}} for details.}
}
\value{
a flextable object
}
\description{
Internal function for formatting flextables
}
