% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Syntax-Functions.R
\name{write_efa}
\alias{write_efa}
\title{Write exploratory factor analysis syntax}
\usage{
write_efa(nf, vnames)
}
\arguments{
\item{nf}{integer; number of factors}

\item{vnames}{character vector; names of variables to include in the efa}
}
\value{
character. Use \code{cat()} to best examine the returned syntax.
}
\description{
Converts variable names to lavaan compatible exploratory factor analysis syntax
}
\examples{
vnames <- paste("x", 1:10)
syntax <- write_efa(nf = 2, vnames = vnames)
cat(syntax)

}
