% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate-Functions.R
\name{agg_cors}
\alias{agg_cors}
\title{Aggregated factor correlations}
\usage{
agg_cors(models, flag = 0.9, type = "factor")
}
\arguments{
\item{models}{An object returned from \code{\link[kfa]{kfa}}}

\item{flag}{threshold above which a factor correlation will be flagged}

\item{type}{currently ignored; \code{"factor"} (default) or \code{"observed"} variable correlations}
}
\value{
\code{data.frame} of mean factor correlations for each factor model and \code{vector} with count of folds with a flagged correlation
}
\description{
The factor correlations aggregated over k-folds
}
\examples{
data(example.kfa)
agg_cors(example.kfa)

}
