% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate-Functions.R
\name{agg_loadings}
\alias{agg_loadings}
\title{Aggregated factor loadings}
\usage{
agg_loadings(models, flag = 0.3, digits = 2)
}
\arguments{
\item{models}{An object returned from \code{\link[kfa]{kfa}}}

\item{flag}{threshold below which loading will be flagged}

\item{digits}{integer; number of decimal places to display in the report.}
}
\value{
\code{data.frame} of mean factor loadings for each factor model and \code{vector} with count of folds with a flagged loading
}
\description{
The factor loadings aggregated over k-folds
}
\examples{
data(example.kfa)
agg_loadings(example.kfa)

}
