% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa_cfa_syntax.R
\name{efa_cfa_syntax}
\alias{efa_cfa_syntax}
\title{Converts EFA results into lavaan CFA syntax}
\usage{
efa_cfa_syntax(
  loadings,
  simple = TRUE,
  threshold = NA,
  single.item = c("keep", "drop", "")
)
}
\description{
Converts EFA results into lavaan CFA syntax
}
