% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findk.R
\name{findk}
\alias{findk}
\title{Determine k for k-fold cross-validation}
\usage{
findk(variables, m, rmsea0 = 0.05, rmseaA = 0.08, ...)
}
\arguments{
\item{m}{integer; maximum number of factors to extract. Default is 4 items per factor.}

\item{rmsea0}{numeric; RMSEA under the null hypothesis.}

\item{rmseaA}{numeric; RMSEA under the alternative hypothesis.}

\item{p}{integer specifying the number of items or \code{data.frame}
where the number of items is determined via \code{\link[base]{ncol}}}
}
\value{
integer indicating the number of folds
}
\description{
This function is specifically for determining k in the context of factor analysis.
Using change in RMSEA as the criterion for obtaining the best fitting factor
analytic model, the function determines the number of folds based on the sample
size, maximum number of possible factors, and RMSEA values.
}
