% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_efa.R
\name{k_efa}
\alias{k_efa}
\title{k-fold exploratory factor analysis}
\usage{
k_efa(variables, m, rotation, ordered, estimator, missing, ...)
}
\arguments{
\item{variables}{a \code{data.frame} of variables to factor analyze.}

\item{m}{an integer; maximum number of factors to extract.}

\item{rotation}{character (case-sensitive); any rotation method listed in
\code{\link[GPArotation]{rotations}} in the \code{GPArotation} package.}

\item{ordered}{passed to \code{lavaan} functions. See \code{\link[lavaan]{lavCor}}.}

\item{estimator}{passed to \code{lavaan} functions. See \code{\link[lavaan]{lavCor}}.}

\item{missing}{passed to \code{lavaan} functions. See \code{\link[lavaan]{lavCor}}.}

\item{...}{other arguments passed to \code{lavaan} functions. See \code{\link[lavaan]{lavOptions}}.}
}
\value{
A list containing \code{lavaan} compatible CFA syntax.
}
\description{
Conducts a sequence of EFAs and converts the resulting factor structure into \code{lavaan} compatible CFA syntax.
}
