% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfa_report.R
\name{kfa_report}
\alias{kfa_report}
\title{Creates summary report from a k-fold factor analysis}
\usage{
kfa_report(
  kfa,
  file.name,
  report.format = "html_document",
  report.title = file.name,
  cut = 0.3
)
}
\arguments{
\item{kfa}{an object returned from \code{kfa} or otherwise defined list of
lists where the inner list consists of \code{lavaan} objects.}

\item{file.name}{character; file name to create on disk.}

\item{report.title}{title of the report}

\item{cut}{numeric; standardized factor loadings below this value are emphasized in
tables and figures.}
}
\value{
html report generated via rmarkdown
}
\description{
Generates a report summarizing the factor analytic results over k-folds.
}
