% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfa_report.R
\name{kfa_report}
\alias{kfa_report}
\title{Creates summary report from a k-fold factor analysis}
\usage{
kfa_report(kfa, instrument.name, file.name = "default")
}
\arguments{
\item{kfa}{an object returned from \code{kfa} or otherwise defined list of
lists where the inner list consists of \code{lavaan} objects.}

\item{instrument.name}{character; name given to the set of variables used in the
factor analysis and printed in the title of the report.}

\item{file.name}{character;  file name to create on disk. The default is
"kfa - \code{instrument.name}".}
}
\value{
html report generated via rmarkdown
}
\description{
Generates a report summarizing the factor analytic results over k-folds.
}
