% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfa_report.R
\name{kfa_report}
\alias{kfa_report}
\title{Creates summary report from a k-fold factor analysis}
\usage{
kfa_report(
  models,
  file.name,
  report.title = file.name,
  report.format = "html_document",
  word.template = NULL,
  index = c("chisq", "cfi", "rmsea"),
  load.flag = 0.3,
  cor.flag = 0.9,
  rel.flag = 0.6,
  digits = 2
)
}
\arguments{
\item{models}{An object returned from \code{\link[kfa]{kfa}}}

\item{file.name}{Character; file name to create on disk.}

\item{report.title}{Title of the report}

\item{report.format}{File format of the report. Default is HTML ("html_document"). See \code{\link[rmarkdown]{render}} for other options.}

\item{word.template}{File path to word document to use as a formatting template when \code{report.format = "word_document"}.}

\item{index}{One or more fit indices to summarize in the report. The degrees of freedom are always reported. Default are "chisq", "cfi", and "rmsea".}

\item{load.flag}{Factor loadings of variables below this value will be flagged. Default is .30}

\item{cor.flag}{Factor correlations above this value will be flagged. Default is .90}

\item{rel.flag}{Factor (scale) reliabilities below this value will be flagged. Default is .60.}

\item{digits}{integer; number of decimal places to display in the report.}
}
\value{
a summary report of factor structures and model fit within and between folds
}
\description{
Generates a report summarizing the factor analytic results over k-folds.
}
