% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate-Functions.R
\name{model_flags}
\alias{model_flags}
\title{Flag model problems}
\usage{
model_flags(kfa, cors, rels, loads)
}
\arguments{
\item{kfa}{An object returned from \code{\link[kfa]{kfa}}}

\item{rels}{An object returned from \code{\link[kfa]{agg_rels}}}

\item{corrs}{An object returned from \code{\link[kfa]{agg_cors}}}
}
\value{
a \code{data.frame}
}
\description{
Internal function to create table of flag counts
}
