% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate-Functions.R
\name{model_structure}
\alias{model_structure}
\title{Unique factor structures}
\usage{
model_structure(kfa, which = c("cfa", "efa"))
}
\arguments{
\item{kfa}{An object returned from \code{\link[kfa]{kfa}}}

\item{which}{Should the unique structures be extracted from the "cfa" \code{lavaan} models or "efa" syntax?}
}
\value{
For each unique structure within each factor model, a \code{list} containing \code{lavaan} syntax specifying the factor structure and the folds where the structure was identified
}
\description{
Extract unique factor structures across the k-folds
}
