% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Syntax-Functions.R
\name{model_structure}
\alias{model_structure}
\title{Unique factor structures}
\usage{
model_structure(models, which = "cfa")
}
\arguments{
\item{models}{An object returned from \code{\link[kfa]{kfa}}}

\item{which}{Should the unique structures be extracted from the "cfa" \code{lavaan} or "efa" syntax?}
}
\value{
\code{lavaan} syntax specifying the unique structures within each factor model
in a \code{data.frame} when \code{which = "cfa"} or, when \code{which = "efa"}, a \code{list}
containing the structure and the folds where the structure was identified
}
\description{
Extract unique factor structures across the k-folds
}
