% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Syntax-Functions.R
\name{model_structure}
\alias{model_structure}
\title{Unique factor structures}
\usage{
model_structure(models)
}
\arguments{
\item{models}{An object returned from \code{\link[kfa]{kfa}}}
}
\value{
\code{data.frame} with the number of folds the unique factor structure was tested for each factor model.
}
\description{
Extract unique factor structures across the k-folds
}
\examples{
data(example.kfa)
model_structure(example.kfa)

}
