% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_efa.R
\name{run_efa}
\alias{run_efa}
\title{Conducts exploratory factor analysis}
\usage{
run_efa(
  items,
  rotation = "oblimin",
  m = floor(ncol(items)/4),
  simple = TRUE,
  threshold = NA,
  ordered = FALSE,
  missing = "listwise",
  ...
)
}
\arguments{
\item{items}{dataframe of item responses}

\item{rotation}{character; any rotation method listed in \code{\link[GPArotation]{rotations}}
in the \code{GPArotation} package. Default is "oblimin".}

\item{m}{integer; maximum number of factors to extract. Default is 4 items per factor.}

\item{simple}{logical; Should the most simple structure be returned (default)?
If \code{FALSE}, items can cross load on multiple factors.}

\item{threshold}{numeric between 0 and 1 indicating the minimum (absolute) value
of the loading for an item on a factor.}

\item{ordered}{logical; Should items be treated as ordinal and the
polychoric correlations used in the factor analysis? When \code{FALSE} (default)
the Pearson correlation matrix is used. A character vector of item names is
also accepted to prompt estimation of the polychoric correlation matrix.}

\item{missing}{passed to \code{lavaan} functions}

\item{...}{other arguments to pass on to \code{lavaan}}
}
\description{
This function is intended for use on independent samples rather than integrated
with k-fold cross-validation.
}
