% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_efa.R
\name{run_efa}
\alias{run_efa}
\title{Conducts exploratory factor analysis}
\usage{
run_efa(
  variables,
  m = floor(ncol(variables)/4),
  rotation = "oblimin",
  simple = TRUE,
  threshold = NA,
  single.item = c("keep", "drop", ""),
  ordered = FALSE,
  estimator = NULL,
  missing = "listwise",
  ...
)
}
\arguments{
\item{variables}{dataframe of variables to factor analyze}

\item{m}{integer; maximum number of factors to extract. Default is 4 items per factor.}

\item{rotation}{character (case-sensitive); any rotation method listed in
\code{\link[GPArotation]{rotations}} in the \code{GPArotation} package.
Default is "oblimin".}

\item{simple}{logical; Should the most simple structure be returned (default)?
If \code{FALSE}, items can cross load on multiple factors.}

\item{threshold}{numeric between 0 and 1 indicating the minimum (absolute) value
of the loading for an item on a factor.}

\item{single.item}{character indicating how single-item factors should be treated.
Use "keep" to keep them in the model when generating the CFA syntax, "drop"
to remove them, or "" indicating the CFA syntax should not be generated for this model.}

\item{ordered}{logical; Should items be treated as ordinal and the
polychoric correlations used in the factor analysis? When \code{FALSE} (default)
the Pearson correlation matrix is used. A character vector of item names is
also accepted to prompt estimation of the polychoric correlation matrix.}

\item{estimator}{passed to \code{lavaan} functions. See \code{\link[lavaan]{lavCor}}.}

\item{missing}{passed to \code{lavaan} functions}

\item{...}{other arguments to pass on to \code{lavaan}}
}
\description{
This function is intended for use on independent samples rather than integrated
with k-fold cross-validation.
}
